Update SLF4J documentation and example
