+++ slf4j/src/test/java/org/slf4j/impl/StaticLoggerBinder.java
@@ -1,0 +1,15 @@
+package org.slf4j.impl;
+
+import io.odin.slf4j.ExternalLogger;
+
+public class StaticLoggerBinder extends ExternalLogger {
+
+    public static String REQUESTED_API_VERSION = "1.7";
+
+    private static final StaticLoggerBinder _instance = new StaticLoggerBinder();
+
+    public static StaticLoggerBinder getSingleton() {
+        return _instance;
+    }
+
+}
