+++ slf4j/src/test/scala/io/odin/slf4j/BufferingLogger.scala
@@ -1,13 +1,13 @@
 package io.odin.slf4j
 
 import cats.effect.concurrent.Ref
-import cats.effect.{Sync, Timer}
-import io.odin.{Level, LoggerMessage}
+import cats.effect.{Clock, Sync}
 import io.odin.loggers.DefaultLogger
+import io.odin.{Level, LoggerMessage}
 
 import scala.collection.immutable.Queue
 
-class BufferingLogger[F[_]: Timer](override val minLevel: Level)(implicit F: Sync[F])
+class BufferingLogger[F[_]: Clock](override val minLevel: Level)(implicit F: Sync[F])
     extends DefaultLogger[F](minLevel) {
 
   val buffer: Ref[F, Queue[LoggerMessage]] = Ref.unsafe[F, Queue[LoggerMessage]](Queue.empty)
